[refactor] move `mod.rs` out of folders
